% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L1000_compounds}
\alias{L1000_compounds}
\title{L1000_compounds}
\format{
A character vector with 1679 elements.
\describe{
  \item{id}{The unique identifier}
  \item{name}{The name associated with the id}
}
}
\source{
\url{http://www.source-of-my-data.com}
}
\usage{
L1000_compounds
}
\description{
Compounds included in 2017 Release TCSs as published in Stathias et al., 2018
}
\keyword{datasets}
