% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L1000_genes}
\alias{L1000_genes}
\title{L1000_genes}
\format{
chr vector 961 elements
\describe{
  \item{id}{The unique identifier}
  \item{name}{The name associated with the id}
}
}
\source{
\url{http://www.source-of-my-data.com}
}
\usage{
L1000_genes
}
\description{
character vector of L1000-measured genes
}
\keyword{datasets}
