% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LINCS.ResponseSigs}
\alias{LINCS.ResponseSigs}
\title{LINCS.ResponseSigs}
\format{
A data frame with 1679 rows and 961 variables:
\describe{
  \item{id}{The unique identifier}
  \item{name}{The name associated with the id}
}
}
\source{
\url{http://www.source-of-my-data.com}
}
\usage{
LINCS.ResponseSigs
}
\description{
2017 Release TCSs as published in Stathias et al., 2018
}
\keyword{datasets}
